requirements.txt
opencv-python==4.8.1.78
hyperlpr3==0.1.3
pillow==9.5.0
ultralytics==8.1.24